requirements.txt
boto3>=1.34.0
openpyxl>=3.1.0
python-dateutil>=2.8.0